# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "媒体同步和刮削系统API",
        "title": "QMediaSync API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8115",
    "basePath": "/",
    "paths": {
        "/115/newurl": {
            "get": {
                "description": "根据pickcode查询115文件直链并按需302跳转",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "115开放平台"
                ],
                "summary": "获取115文件直链",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件PickCode",
                        "name": "pickcode",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "115用户ID",
                        "name": "userid",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "是否强制直链播放，1为直链，0使用本地代理时会走代理",
                        "name": "force",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "302": {
                        "description": "重定向到文件直链",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/auth/115-qrcode-open": {
            "post": {
                "security": [
                    {
                        "JwtAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "生成115开放平台登录二维码并异步轮询状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "115开放平台"
                ],
                "summary": "获取115登录二维码",
                "parameters": [
                    {
                        "description": "账号ID",
                        "name": "account_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/auth/115-qrcode-status": {
            "post": {
                "security": [
                    {
                        "JwtAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "查询指定二维码UID的扫码进度",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "115开放平台"
                ],
                "summary": "查询115二维码扫码状态",
                "parameters": [
                    {
                        "description": "二维码UID",
                        "name": "uid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "账号ID",
                        "name": "account_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/auth/115-status": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取指定115账号的登录状态及存储信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "115开放平台"
                ],
                "summary": "查询115账号状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账号ID",
                        "name": "account_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "api_key",
            "in": "query"
        },
        "JwtAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}